fix Story model
